make message legacy
