+++ proto/zetachain/zetacore/crosschain/tx.proto
@@ -38,12 +38,6 @@ service Msg {
 
   rpc UpdateRateLimiterFlags(MsgUpdateRateLimiterFlags)
       returns (MsgUpdateRateLimiterFlagsResponse);
-
-  rpc MigrateERC20CustodyFunds(MsgMigrateERC20CustodyFunds)
-      returns (MsgMigrateERC20CustodyFundsResponse);
-
-  rpc UpdateERC20CustodyPauseStatus(MsgUpdateERC20CustodyPauseStatus)
-      returns (MsgUpdateERC20CustodyPauseStatusResponse);
 }
 
 message MsgMigrateTssFunds {
@@ -252,29 +246,3 @@ message MsgUpdateRateLimiterFlags {
 }
 
 message MsgUpdateRateLimiterFlagsResponse {}
-
-message MsgMigrateERC20CustodyFunds {
-  option (cosmos.msg.v1.signer) = "creator";
-  string creator = 1;
-  int64 chain_id = 2;
-  string new_custody_address = 3;
-  string erc20_address = 4;
-  string amount = 5 [
-    (gogoproto.customtype) = "cosmossdk.io/math.Uint",
-    (gogoproto.nullable) = false
-  ];
-}
-
-message MsgMigrateERC20CustodyFundsResponse { string cctx_index = 1; }
-
-message MsgUpdateERC20CustodyPauseStatus {
-  option (cosmos.msg.v1.signer) = "creator";
-  string creator = 1;
-  int64 chain_id = 2;
-
-  // pause or unpause
-  // true = pause, false = unpause
-  bool pause = 3;
-}
-
-message MsgUpdateERC20CustodyPauseStatusResponse { string cctx_index = 1; }
