+++ typescript/zetachain/zetacore/crosschain/tx_pb.ts
@@ -22,7 +22,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file zetachain/zetacore/crosschain/tx.proto.
  */
 export const file_zetachain_zetacore_crosschain_tx: GenFile = /*@__PURE__*/
-  fileDesc("CiZ6ZXRhY2hhaW4vemV0YWNvcmUvY3Jvc3NjaGFpbi90eC5wcm90bxIdemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4idQoSTXNnTWlncmF0ZVRzc0Z1bmRzEg8KB2NyZWF0b3IYASABKAkSEAoIY2hhaW5faWQYAiABKAMSLgoGYW1vdW50GAMgASgJQh7I3h8A2t4fFmNvc21vc3Nkay5pby9tYXRoLlVpbnQ6DILnsCoHY3JlYXRvciIcChpNc2dNaWdyYXRlVHNzRnVuZHNSZXNwb25zZSJIChNNc2dVcGRhdGVUc3NBZGRyZXNzEg8KB2NyZWF0b3IYASABKAkSEgoKdHNzX3B1YmtleRgCIAEoCToMguewKgdjcmVhdG9yIh0KG01zZ1VwZGF0ZVRzc0FkZHJlc3NSZXNwb25zZSL5AQoUTXNnQWRkSW5ib3VuZFRyYWNrZXISDwoHY3JlYXRvchgBIAEoCRIQCghjaGFpbl9pZBgCIAEoAxIPCgd0eF9oYXNoGAMgASgJEjgKCWNvaW5fdHlwZRgEIAEoDjIlLnpldGFjaGFpbi56ZXRhY29yZS5wa2cuY29pbi5Db2luVHlwZRI3CgVwcm9vZhgFIAEoCzIkLnpldGFjaGFpbi56ZXRhY29yZS5wa2cucHJvb2ZzLlByb29mQgIYARIWCgpibG9ja19oYXNoGAYgASgJQgIYARIUCgh0eF9pbmRleBgHIAEoA0ICGAE6DILnsCoHY3JlYXRvciIeChxNc2dBZGRJbmJvdW5kVHJhY2tlclJlc3BvbnNlIlsKF01zZ1JlbW92ZUluYm91bmRUcmFja2VyEg8KB2NyZWF0b3IYASABKAkSEAoIY2hhaW5faWQYAiABKAMSDwoHdHhfaGFzaBgDIAEoCToMguewKgdjcmVhdG9yIiEKH01zZ1JlbW92ZUluYm91bmRUcmFja2VyUmVzcG9uc2Ui1QEKEU1zZ1doaXRlbGlzdEVSQzIwEg8KB2NyZWF0b3IYASABKAkSFQoNZXJjMjBfYWRkcmVzcxgCIAEoCRIQCghjaGFpbl9pZBgDIAEoAxIMCgRuYW1lGAQgASgJEg4KBnN5bWJvbBgFIAEoCRIQCghkZWNpbWFscxgGIAEoDRIRCglnYXNfbGltaXQYByABKAMSNQoNbGlxdWlkaXR5X2NhcBgIIAEoCUIeyN4fANreHxZjb3Ntb3NzZGsuaW8vbWF0aC5VaW50OgyC57AqB2NyZWF0b3IiRgoZTXNnV2hpdGVsaXN0RVJDMjBSZXNwb25zZRIVCg16cmMyMF9hZGRyZXNzGAEgASgJEhIKCmNjdHhfaW5kZXgYAiABKAkizwEKFU1zZ0FkZE91dGJvdW5kVHJhY2tlchIPCgdjcmVhdG9yGAEgASgJEhAKCGNoYWluX2lkGAIgASgDEg0KBW5vbmNlGAMgASgEEg8KB3R4X2hhc2gYBCABKAkSNwoFcHJvb2YYBSABKAsyJC56ZXRhY2hhaW4uemV0YWNvcmUucGtnLnByb29mcy5Qcm9vZkICGAESFgoKYmxvY2tfaGFzaBgGIAEoCUICGAESFAoIdHhfaW5kZXgYByABKANCAhgBOgyC57AqB2NyZWF0b3IiMwodTXNnQWRkT3V0Ym91bmRUcmFja2VyUmVzcG9uc2USEgoKaXNfcmVtb3ZlZBgBIAEoCCJaChhNc2dSZW1vdmVPdXRib3VuZFRyYWNrZXISDwoHY3JlYXRvchgBIAEoCRIQCghjaGFpbl9pZBgCIAEoAxINCgVub25jZRgDIAEoBDoMguewKgdjcmVhdG9yIiIKIE1zZ1JlbW92ZU91dGJvdW5kVHJhY2tlclJlc3BvbnNlIpEBCg9Nc2dWb3RlR2FzUHJpY2USDwoHY3JlYXRvchgBIAEoCRIQCghjaGFpbl9pZBgCIAEoAxINCgVwcmljZRgDIAEoBBIUCgxwcmlvcml0eV9mZWUYBiABKAQSFAoMYmxvY2tfbnVtYmVyGAQgASgEEhIKBnN1cHBseRgFIAEoCUICGAE6DILnsCoHY3JlYXRvciIZChdNc2dWb3RlR2FzUHJpY2VSZXNwb25zZSL1BAoPTXNnVm90ZU91dGJvdW5kEg8KB2NyZWF0b3IYASABKAkSEQoJY2N0eF9oYXNoGAIgASgJEh4KFm9ic2VydmVkX291dGJvdW5kX2hhc2gYAyABKAkSJgoeb2JzZXJ2ZWRfb3V0Ym91bmRfYmxvY2tfaGVpZ2h0GAQgASgEEiIKGm9ic2VydmVkX291dGJvdW5kX2dhc191c2VkGAogASgEEkwKJW9ic2VydmVkX291dGJvdW5kX2VmZmVjdGl2ZV9nYXNfcHJpY2UYCyABKAlCHcjeHwDa3h8VY29zbW9zc2RrLmlvL21hdGguSW50Ei0KJW9ic2VydmVkX291dGJvdW5kX2VmZmVjdGl2ZV9nYXNfbGltaXQYDCABKAQSTwoOdmFsdWVfcmVjZWl2ZWQYBSABKAlCN8jeHwDa3h8WY29zbW9zc2RrLmlvL21hdGguVWludPLeHxV5YW1sOiJ2YWx1ZV9yZWNlaXZlZCISPAoGc3RhdHVzGAYgASgOMiwuemV0YWNoYWluLnpldGFjb3JlLnBrZy5jaGFpbnMuUmVjZWl2ZVN0YXR1cxIWCg5vdXRib3VuZF9jaGFpbhgHIAEoAxIaChJvdXRib3VuZF90c3Nfbm9uY2UYCCABKAQSOAoJY29pbl90eXBlGAkgASgOMiUuemV0YWNoYWluLnpldGFjb3JlLnBrZy5jb2luLkNvaW5UeXBlEkoKEWNvbmZpcm1hdGlvbl9tb2RlGA0gASgOMi8uemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uQ29uZmlybWF0aW9uTW9kZToMguewKgdjcmVhdG9yIhkKF01zZ1ZvdGVPdXRib3VuZFJlc3BvbnNlIosGCg5Nc2dWb3RlSW5ib3VuZBIPCgdjcmVhdG9yGAEgASgJEg4KBnNlbmRlchgCIAEoCRIXCg9zZW5kZXJfY2hhaW5faWQYAyABKAMSEAoIcmVjZWl2ZXIYBCABKAkSFgoOcmVjZWl2ZXJfY2hhaW4YBSABKAMSLgoGYW1vdW50GAYgASgJQh7I3h8A2t4fFmNvc21vc3Nkay5pby9tYXRoLlVpbnQSDwoHbWVzc2FnZRgIIAEoCRIUCgxpbmJvdW5kX2hhc2gYCSABKAkSHAoUaW5ib3VuZF9ibG9ja19oZWlnaHQYCiABKAQSEQoJZ2FzX2xpbWl0GAsgASgEEjgKCWNvaW5fdHlwZRgMIAEoDjIlLnpldGFjaGFpbi56ZXRhY29yZS5wa2cuY29pbi5Db2luVHlwZRIRCgl0eF9vcmlnaW4YDSABKAkSDQoFYXNzZXQYDiABKAkSEwoLZXZlbnRfaW5kZXgYDyABKAQSWQoZcHJvdG9jb2xfY29udHJhY3RfdmVyc2lvbhgQIAEoDjI2LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLlByb3RvY29sQ29udHJhY3RWZXJzaW9uEkoKDnJldmVydF9vcHRpb25zGBEgASgLMiwuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uUmV2ZXJ0T3B0aW9uc0IEyN4fABJACgxjYWxsX29wdGlvbnMYEiABKAsyKi56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5DYWxsT3B0aW9ucxIbChNpc19jcm9zc19jaGFpbl9jYWxsGBMgASgIEjwKBnN0YXR1cxgUIAEoDjIsLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLkluYm91bmRTdGF0dXMSSgoRY29uZmlybWF0aW9uX21vZGUYFSABKA4yLy56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Db25maXJtYXRpb25Nb2RlOgyC57AqB2NyZWF0b3IiGAoWTXNnVm90ZUluYm91bmRSZXNwb25zZSJGChFNc2dBYm9ydFN0dWNrQ0NUWBIPCgdjcmVhdG9yGAEgASgJEhIKCmNjdHhfaW5kZXgYAiABKAk6DILnsCoHY3JlYXRvciIbChlNc2dBYm9ydFN0dWNrQ0NUWFJlc3BvbnNlImEKFE1zZ1JlZnVuZEFib3J0ZWRDQ1RYEg8KB2NyZWF0b3IYASABKAkSEgoKY2N0eF9pbmRleBgCIAEoCRIWCg5yZWZ1bmRfYWRkcmVzcxgDIAEoCToMguewKgdjcmVhdG9yIh4KHE1zZ1JlZnVuZEFib3J0ZWRDQ1RYUmVzcG9uc2UijQEKGU1zZ1VwZGF0ZVJhdGVMaW1pdGVyRmxhZ3MSDwoHY3JlYXRvchgBIAEoCRJRChJyYXRlX2xpbWl0ZXJfZmxhZ3MYAiABKAsyLy56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5SYXRlTGltaXRlckZsYWdzQgTI3h8AOgyC57AqB2NyZWF0b3IiIwohTXNnVXBkYXRlUmF0ZUxpbWl0ZXJGbGFnc1Jlc3BvbnNlIrIBChtNc2dNaWdyYXRlRVJDMjBDdXN0b2R5RnVuZHMSDwoHY3JlYXRvchgBIAEoCRIQCghjaGFpbl9pZBgCIAEoAxIbChNuZXdfY3VzdG9keV9hZGRyZXNzGAMgASgJEhUKDWVyYzIwX2FkZHJlc3MYBCABKAkSLgoGYW1vdW50GAUgASgJQh7I3h8A2t4fFmNvc21vc3Nkay5pby9tYXRoLlVpbnQ6DILnsCoHY3JlYXRvciI5CiNNc2dNaWdyYXRlRVJDMjBDdXN0b2R5RnVuZHNSZXNwb25zZRISCgpjY3R4X2luZGV4GAEgASgJImIKIE1zZ1VwZGF0ZUVSQzIwQ3VzdG9keVBhdXNlU3RhdHVzEg8KB2NyZWF0b3IYASABKAkSEAoIY2hhaW5faWQYAiABKAMSDQoFcGF1c2UYAyABKAg6DILnsCoHY3JlYXRvciI+CihNc2dVcGRhdGVFUkMyMEN1c3RvZHlQYXVzZVN0YXR1c1Jlc3BvbnNlEhIKCmNjdHhfaW5kZXgYASABKAkykxAKA01zZxKIAQoSQWRkT3V0Ym91bmRUcmFja2VyEjQuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnQWRkT3V0Ym91bmRUcmFja2VyGjwuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnQWRkT3V0Ym91bmRUcmFja2VyUmVzcG9uc2UShQEKEUFkZEluYm91bmRUcmFja2VyEjMuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnQWRkSW5ib3VuZFRyYWNrZXIaOy56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dBZGRJbmJvdW5kVHJhY2tlclJlc3BvbnNlEo4BChRSZW1vdmVJbmJvdW5kVHJhY2tlchI2LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1JlbW92ZUluYm91bmRUcmFja2VyGj4uemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnUmVtb3ZlSW5ib3VuZFRyYWNrZXJSZXNwb25zZRKRAQoVUmVtb3ZlT3V0Ym91bmRUcmFja2VyEjcuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnUmVtb3ZlT3V0Ym91bmRUcmFja2VyGj8uemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnUmVtb3ZlT3V0Ym91bmRUcmFja2VyUmVzcG9uc2USdgoMVm90ZUdhc1ByaWNlEi4uemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVm90ZUdhc1ByaWNlGjYuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVm90ZUdhc1ByaWNlUmVzcG9uc2USdgoMVm90ZU91dGJvdW5kEi4uemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVm90ZU91dGJvdW5kGjYuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVm90ZU91dGJvdW5kUmVzcG9uc2UScwoLVm90ZUluYm91bmQSLS56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dWb3RlSW5ib3VuZBo1LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1ZvdGVJbmJvdW5kUmVzcG9uc2USfAoOV2hpdGVsaXN0RVJDMjASMC56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dXaGl0ZWxpc3RFUkMyMBo4LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1doaXRlbGlzdEVSQzIwUmVzcG9uc2USggEKEFVwZGF0ZVRzc0FkZHJlc3MSMi56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dVcGRhdGVUc3NBZGRyZXNzGjouemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVXBkYXRlVHNzQWRkcmVzc1Jlc3BvbnNlEn8KD01pZ3JhdGVUc3NGdW5kcxIxLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ01pZ3JhdGVUc3NGdW5kcxo5LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ01pZ3JhdGVUc3NGdW5kc1Jlc3BvbnNlEnwKDkFib3J0U3R1Y2tDQ1RYEjAuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnQWJvcnRTdHVja0NDVFgaOC56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dBYm9ydFN0dWNrQ0NUWFJlc3BvbnNlEoUBChFSZWZ1bmRBYm9ydGVkQ0NUWBIzLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1JlZnVuZEFib3J0ZWRDQ1RYGjsuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnUmVmdW5kQWJvcnRlZENDVFhSZXNwb25zZRKUAQoWVXBkYXRlUmF0ZUxpbWl0ZXJGbGFncxI4LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1VwZGF0ZVJhdGVMaW1pdGVyRmxhZ3MaQC56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dVcGRhdGVSYXRlTGltaXRlckZsYWdzUmVzcG9uc2USmgEKGE1pZ3JhdGVFUkMyMEN1c3RvZHlGdW5kcxI6LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ01pZ3JhdGVFUkMyMEN1c3RvZHlGdW5kcxpCLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ01pZ3JhdGVFUkMyMEN1c3RvZHlGdW5kc1Jlc3BvbnNlEqkBCh1VcGRhdGVFUkMyMEN1c3RvZHlQYXVzZVN0YXR1cxI/LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1VwZGF0ZUVSQzIwQ3VzdG9keVBhdXNlU3RhdHVzGkcuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVXBkYXRlRVJDMjBDdXN0b2R5UGF1c2VTdGF0dXNSZXNwb25zZRoFgOewKgFC8QEKIWNvbS56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbkIHVHhQcm90b1ABWi1naXRodWIuY29tL3pldGEtY2hhaW4vbm9kZS94L2Nyb3NzY2hhaW4vdHlwZXOiAgNaWkOqAh1aZXRhY2hhaW4uWmV0YWNvcmUuQ3Jvc3NjaGFpbsoCHVpldGFjaGFpblxaZXRhY29yZVxDcm9zc2NoYWlu4gIpWmV0YWNoYWluXFpldGFjb3JlXENyb3NzY2hhaW5cR1BCTWV0YWRhdGHqAh9aZXRhY2hhaW46OlpldGFjb3JlOjpDcm9zc2NoYWluYgZwcm90bzM", [file_gogoproto_gogo, file_zetachain_zetacore_pkg_chains_chains, file_zetachain_zetacore_pkg_coin_coin, file_zetachain_zetacore_pkg_proofs_proofs, file_zetachain_zetacore_crosschain_rate_limiter_flags, file_zetachain_zetacore_crosschain_cross_chain_tx, file_cosmos_msg_v1_msg]);
+  fileDesc("CiZ6ZXRhY2hhaW4vemV0YWNvcmUvY3Jvc3NjaGFpbi90eC5wcm90bxIdemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4idQoSTXNnTWlncmF0ZVRzc0Z1bmRzEg8KB2NyZWF0b3IYASABKAkSEAoIY2hhaW5faWQYAiABKAMSLgoGYW1vdW50GAMgASgJQh7I3h8A2t4fFmNvc21vc3Nkay5pby9tYXRoLlVpbnQ6DILnsCoHY3JlYXRvciIcChpNc2dNaWdyYXRlVHNzRnVuZHNSZXNwb25zZSJIChNNc2dVcGRhdGVUc3NBZGRyZXNzEg8KB2NyZWF0b3IYASABKAkSEgoKdHNzX3B1YmtleRgCIAEoCToMguewKgdjcmVhdG9yIh0KG01zZ1VwZGF0ZVRzc0FkZHJlc3NSZXNwb25zZSL5AQoUTXNnQWRkSW5ib3VuZFRyYWNrZXISDwoHY3JlYXRvchgBIAEoCRIQCghjaGFpbl9pZBgCIAEoAxIPCgd0eF9oYXNoGAMgASgJEjgKCWNvaW5fdHlwZRgEIAEoDjIlLnpldGFjaGFpbi56ZXRhY29yZS5wa2cuY29pbi5Db2luVHlwZRI3CgVwcm9vZhgFIAEoCzIkLnpldGFjaGFpbi56ZXRhY29yZS5wa2cucHJvb2ZzLlByb29mQgIYARIWCgpibG9ja19oYXNoGAYgASgJQgIYARIUCgh0eF9pbmRleBgHIAEoA0ICGAE6DILnsCoHY3JlYXRvciIeChxNc2dBZGRJbmJvdW5kVHJhY2tlclJlc3BvbnNlIlsKF01zZ1JlbW92ZUluYm91bmRUcmFja2VyEg8KB2NyZWF0b3IYASABKAkSEAoIY2hhaW5faWQYAiABKAMSDwoHdHhfaGFzaBgDIAEoCToMguewKgdjcmVhdG9yIiEKH01zZ1JlbW92ZUluYm91bmRUcmFja2VyUmVzcG9uc2Ui1QEKEU1zZ1doaXRlbGlzdEVSQzIwEg8KB2NyZWF0b3IYASABKAkSFQoNZXJjMjBfYWRkcmVzcxgCIAEoCRIQCghjaGFpbl9pZBgDIAEoAxIMCgRuYW1lGAQgASgJEg4KBnN5bWJvbBgFIAEoCRIQCghkZWNpbWFscxgGIAEoDRIRCglnYXNfbGltaXQYByABKAMSNQoNbGlxdWlkaXR5X2NhcBgIIAEoCUIeyN4fANreHxZjb3Ntb3NzZGsuaW8vbWF0aC5VaW50OgyC57AqB2NyZWF0b3IiRgoZTXNnV2hpdGVsaXN0RVJDMjBSZXNwb25zZRIVCg16cmMyMF9hZGRyZXNzGAEgASgJEhIKCmNjdHhfaW5kZXgYAiABKAkizwEKFU1zZ0FkZE91dGJvdW5kVHJhY2tlchIPCgdjcmVhdG9yGAEgASgJEhAKCGNoYWluX2lkGAIgASgDEg0KBW5vbmNlGAMgASgEEg8KB3R4X2hhc2gYBCABKAkSNwoFcHJvb2YYBSABKAsyJC56ZXRhY2hhaW4uemV0YWNvcmUucGtnLnByb29mcy5Qcm9vZkICGAESFgoKYmxvY2tfaGFzaBgGIAEoCUICGAESFAoIdHhfaW5kZXgYByABKANCAhgBOgyC57AqB2NyZWF0b3IiMwodTXNnQWRkT3V0Ym91bmRUcmFja2VyUmVzcG9uc2USEgoKaXNfcmVtb3ZlZBgBIAEoCCJaChhNc2dSZW1vdmVPdXRib3VuZFRyYWNrZXISDwoHY3JlYXRvchgBIAEoCRIQCghjaGFpbl9pZBgCIAEoAxINCgVub25jZRgDIAEoBDoMguewKgdjcmVhdG9yIiIKIE1zZ1JlbW92ZU91dGJvdW5kVHJhY2tlclJlc3BvbnNlIpEBCg9Nc2dWb3RlR2FzUHJpY2USDwoHY3JlYXRvchgBIAEoCRIQCghjaGFpbl9pZBgCIAEoAxINCgVwcmljZRgDIAEoBBIUCgxwcmlvcml0eV9mZWUYBiABKAQSFAoMYmxvY2tfbnVtYmVyGAQgASgEEhIKBnN1cHBseRgFIAEoCUICGAE6DILnsCoHY3JlYXRvciIZChdNc2dWb3RlR2FzUHJpY2VSZXNwb25zZSL1BAoPTXNnVm90ZU91dGJvdW5kEg8KB2NyZWF0b3IYASABKAkSEQoJY2N0eF9oYXNoGAIgASgJEh4KFm9ic2VydmVkX291dGJvdW5kX2hhc2gYAyABKAkSJgoeb2JzZXJ2ZWRfb3V0Ym91bmRfYmxvY2tfaGVpZ2h0GAQgASgEEiIKGm9ic2VydmVkX291dGJvdW5kX2dhc191c2VkGAogASgEEkwKJW9ic2VydmVkX291dGJvdW5kX2VmZmVjdGl2ZV9nYXNfcHJpY2UYCyABKAlCHcjeHwDa3h8VY29zbW9zc2RrLmlvL21hdGguSW50Ei0KJW9ic2VydmVkX291dGJvdW5kX2VmZmVjdGl2ZV9nYXNfbGltaXQYDCABKAQSTwoOdmFsdWVfcmVjZWl2ZWQYBSABKAlCN8jeHwDa3h8WY29zbW9zc2RrLmlvL21hdGguVWludPLeHxV5YW1sOiJ2YWx1ZV9yZWNlaXZlZCISPAoGc3RhdHVzGAYgASgOMiwuemV0YWNoYWluLnpldGFjb3JlLnBrZy5jaGFpbnMuUmVjZWl2ZVN0YXR1cxIWCg5vdXRib3VuZF9jaGFpbhgHIAEoAxIaChJvdXRib3VuZF90c3Nfbm9uY2UYCCABKAQSOAoJY29pbl90eXBlGAkgASgOMiUuemV0YWNoYWluLnpldGFjb3JlLnBrZy5jb2luLkNvaW5UeXBlEkoKEWNvbmZpcm1hdGlvbl9tb2RlGA0gASgOMi8uemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uQ29uZmlybWF0aW9uTW9kZToMguewKgdjcmVhdG9yIhkKF01zZ1ZvdGVPdXRib3VuZFJlc3BvbnNlIosGCg5Nc2dWb3RlSW5ib3VuZBIPCgdjcmVhdG9yGAEgASgJEg4KBnNlbmRlchgCIAEoCRIXCg9zZW5kZXJfY2hhaW5faWQYAyABKAMSEAoIcmVjZWl2ZXIYBCABKAkSFgoOcmVjZWl2ZXJfY2hhaW4YBSABKAMSLgoGYW1vdW50GAYgASgJQh7I3h8A2t4fFmNvc21vc3Nkay5pby9tYXRoLlVpbnQSDwoHbWVzc2FnZRgIIAEoCRIUCgxpbmJvdW5kX2hhc2gYCSABKAkSHAoUaW5ib3VuZF9ibG9ja19oZWlnaHQYCiABKAQSEQoJZ2FzX2xpbWl0GAsgASgEEjgKCWNvaW5fdHlwZRgMIAEoDjIlLnpldGFjaGFpbi56ZXRhY29yZS5wa2cuY29pbi5Db2luVHlwZRIRCgl0eF9vcmlnaW4YDSABKAkSDQoFYXNzZXQYDiABKAkSEwoLZXZlbnRfaW5kZXgYDyABKAQSWQoZcHJvdG9jb2xfY29udHJhY3RfdmVyc2lvbhgQIAEoDjI2LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLlByb3RvY29sQ29udHJhY3RWZXJzaW9uEkoKDnJldmVydF9vcHRpb25zGBEgASgLMiwuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uUmV2ZXJ0T3B0aW9uc0IEyN4fABJACgxjYWxsX29wdGlvbnMYEiABKAsyKi56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5DYWxsT3B0aW9ucxIbChNpc19jcm9zc19jaGFpbl9jYWxsGBMgASgIEjwKBnN0YXR1cxgUIAEoDjIsLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLkluYm91bmRTdGF0dXMSSgoRY29uZmlybWF0aW9uX21vZGUYFSABKA4yLy56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Db25maXJtYXRpb25Nb2RlOgyC57AqB2NyZWF0b3IiGAoWTXNnVm90ZUluYm91bmRSZXNwb25zZSJGChFNc2dBYm9ydFN0dWNrQ0NUWBIPCgdjcmVhdG9yGAEgASgJEhIKCmNjdHhfaW5kZXgYAiABKAk6DILnsCoHY3JlYXRvciIbChlNc2dBYm9ydFN0dWNrQ0NUWFJlc3BvbnNlImEKFE1zZ1JlZnVuZEFib3J0ZWRDQ1RYEg8KB2NyZWF0b3IYASABKAkSEgoKY2N0eF9pbmRleBgCIAEoCRIWCg5yZWZ1bmRfYWRkcmVzcxgDIAEoCToMguewKgdjcmVhdG9yIh4KHE1zZ1JlZnVuZEFib3J0ZWRDQ1RYUmVzcG9uc2UijQEKGU1zZ1VwZGF0ZVJhdGVMaW1pdGVyRmxhZ3MSDwoHY3JlYXRvchgBIAEoCRJRChJyYXRlX2xpbWl0ZXJfZmxhZ3MYAiABKAsyLy56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5SYXRlTGltaXRlckZsYWdzQgTI3h8AOgyC57AqB2NyZWF0b3IiIwohTXNnVXBkYXRlUmF0ZUxpbWl0ZXJGbGFnc1Jlc3BvbnNlMsoNCgNNc2cSiAEKEkFkZE91dGJvdW5kVHJhY2tlchI0LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ0FkZE91dGJvdW5kVHJhY2tlcho8LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ0FkZE91dGJvdW5kVHJhY2tlclJlc3BvbnNlEoUBChFBZGRJbmJvdW5kVHJhY2tlchIzLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ0FkZEluYm91bmRUcmFja2VyGjsuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnQWRkSW5ib3VuZFRyYWNrZXJSZXNwb25zZRKOAQoUUmVtb3ZlSW5ib3VuZFRyYWNrZXISNi56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dSZW1vdmVJbmJvdW5kVHJhY2tlcho+LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1JlbW92ZUluYm91bmRUcmFja2VyUmVzcG9uc2USkQEKFVJlbW92ZU91dGJvdW5kVHJhY2tlchI3LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1JlbW92ZU91dGJvdW5kVHJhY2tlcho/LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1JlbW92ZU91dGJvdW5kVHJhY2tlclJlc3BvbnNlEnYKDFZvdGVHYXNQcmljZRIuLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1ZvdGVHYXNQcmljZRo2LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1ZvdGVHYXNQcmljZVJlc3BvbnNlEnYKDFZvdGVPdXRib3VuZBIuLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1ZvdGVPdXRib3VuZBo2LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1ZvdGVPdXRib3VuZFJlc3BvbnNlEnMKC1ZvdGVJbmJvdW5kEi0uemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVm90ZUluYm91bmQaNS56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dWb3RlSW5ib3VuZFJlc3BvbnNlEnwKDldoaXRlbGlzdEVSQzIwEjAuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnV2hpdGVsaXN0RVJDMjAaOC56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dXaGl0ZWxpc3RFUkMyMFJlc3BvbnNlEoIBChBVcGRhdGVUc3NBZGRyZXNzEjIuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVXBkYXRlVHNzQWRkcmVzcxo6LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1VwZGF0ZVRzc0FkZHJlc3NSZXNwb25zZRJ/Cg9NaWdyYXRlVHNzRnVuZHMSMS56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dNaWdyYXRlVHNzRnVuZHMaOS56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dNaWdyYXRlVHNzRnVuZHNSZXNwb25zZRJ8Cg5BYm9ydFN0dWNrQ0NUWBIwLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ0Fib3J0U3R1Y2tDQ1RYGjguemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnQWJvcnRTdHVja0NDVFhSZXNwb25zZRKFAQoRUmVmdW5kQWJvcnRlZENDVFgSMy56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dSZWZ1bmRBYm9ydGVkQ0NUWBo7LnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluLk1zZ1JlZnVuZEFib3J0ZWRDQ1RYUmVzcG9uc2USlAEKFlVwZGF0ZVJhdGVMaW1pdGVyRmxhZ3MSOC56ZXRhY2hhaW4uemV0YWNvcmUuY3Jvc3NjaGFpbi5Nc2dVcGRhdGVSYXRlTGltaXRlckZsYWdzGkAuemV0YWNoYWluLnpldGFjb3JlLmNyb3NzY2hhaW4uTXNnVXBkYXRlUmF0ZUxpbWl0ZXJGbGFnc1Jlc3BvbnNlGgWA57AqAULxAQohY29tLnpldGFjaGFpbi56ZXRhY29yZS5jcm9zc2NoYWluQgdUeFByb3RvUAFaLWdpdGh1Yi5jb20vemV0YS1jaGFpbi9ub2RlL3gvY3Jvc3NjaGFpbi90eXBlc6ICA1paQ6oCHVpldGFjaGFpbi5aZXRhY29yZS5Dcm9zc2NoYWluygIdWmV0YWNoYWluXFpldGFjb3JlXENyb3NzY2hhaW7iAilaZXRhY2hhaW5cWmV0YWNvcmVcQ3Jvc3NjaGFpblxHUEJNZXRhZGF0YeoCH1pldGFjaGFpbjo6WmV0YWNvcmU6OkNyb3NzY2hhaW5iBnByb3RvMw", [file_gogoproto_gogo, file_zetachain_zetacore_pkg_chains_chains, file_zetachain_zetacore_pkg_coin_coin, file_zetachain_zetacore_pkg_proofs_proofs, file_zetachain_zetacore_crosschain_rate_limiter_flags, file_zetachain_zetacore_crosschain_cross_chain_tx, file_cosmos_msg_v1_msg]);
 
 /**
  * @generated from message zetachain.zetacore.crosschain.MsgMigrateTssFunds
@@ -797,107 +797,6 @@ export const MsgUpdateRateLimiterFlagsResponseSchema: GenMessage<MsgUpdateRateLi
   messageDesc(file_zetachain_zetacore_crosschain_tx, 25);
 
 /**
- * @generated from message zetachain.zetacore.crosschain.MsgMigrateERC20CustodyFunds
- */
-export type MsgMigrateERC20CustodyFunds = Message<"zetachain.zetacore.crosschain.MsgMigrateERC20CustodyFunds"> & {
-  /**
-   * @generated from field: string creator = 1;
-   */
-  creator: string;
-
-  /**
-   * @generated from field: int64 chain_id = 2;
-   */
-  chainId: bigint;
-
-  /**
-   * @generated from field: string new_custody_address = 3;
-   */
-  newCustodyAddress: string;
-
-  /**
-   * @generated from field: string erc20_address = 4;
-   */
-  erc20Address: string;
-
-  /**
-   * @generated from field: string amount = 5;
-   */
-  amount: string;
-};
-
-/**
- * Describes the message zetachain.zetacore.crosschain.MsgMigrateERC20CustodyFunds.
- * Use `create(MsgMigrateERC20CustodyFundsSchema)` to create a new message.
- */
-export const MsgMigrateERC20CustodyFundsSchema: GenMessage<MsgMigrateERC20CustodyFunds> = /*@__PURE__*/
-  messageDesc(file_zetachain_zetacore_crosschain_tx, 26);
-
-/**
- * @generated from message zetachain.zetacore.crosschain.MsgMigrateERC20CustodyFundsResponse
- */
-export type MsgMigrateERC20CustodyFundsResponse = Message<"zetachain.zetacore.crosschain.MsgMigrateERC20CustodyFundsResponse"> & {
-  /**
-   * @generated from field: string cctx_index = 1;
-   */
-  cctxIndex: string;
-};
-
-/**
- * Describes the message zetachain.zetacore.crosschain.MsgMigrateERC20CustodyFundsResponse.
- * Use `create(MsgMigrateERC20CustodyFundsResponseSchema)` to create a new message.
- */
-export const MsgMigrateERC20CustodyFundsResponseSchema: GenMessage<MsgMigrateERC20CustodyFundsResponse> = /*@__PURE__*/
-  messageDesc(file_zetachain_zetacore_crosschain_tx, 27);
-
-/**
- * @generated from message zetachain.zetacore.crosschain.MsgUpdateERC20CustodyPauseStatus
- */
-export type MsgUpdateERC20CustodyPauseStatus = Message<"zetachain.zetacore.crosschain.MsgUpdateERC20CustodyPauseStatus"> & {
-  /**
-   * @generated from field: string creator = 1;
-   */
-  creator: string;
-
-  /**
-   * @generated from field: int64 chain_id = 2;
-   */
-  chainId: bigint;
-
-  /**
-   * pause or unpause
-   * true = pause, false = unpause
-   *
-   * @generated from field: bool pause = 3;
-   */
-  pause: boolean;
-};
-
-/**
- * Describes the message zetachain.zetacore.crosschain.MsgUpdateERC20CustodyPauseStatus.
- * Use `create(MsgUpdateERC20CustodyPauseStatusSchema)` to create a new message.
- */
-export const MsgUpdateERC20CustodyPauseStatusSchema: GenMessage<MsgUpdateERC20CustodyPauseStatus> = /*@__PURE__*/
-  messageDesc(file_zetachain_zetacore_crosschain_tx, 28);
-
-/**
- * @generated from message zetachain.zetacore.crosschain.MsgUpdateERC20CustodyPauseStatusResponse
- */
-export type MsgUpdateERC20CustodyPauseStatusResponse = Message<"zetachain.zetacore.crosschain.MsgUpdateERC20CustodyPauseStatusResponse"> & {
-  /**
-   * @generated from field: string cctx_index = 1;
-   */
-  cctxIndex: string;
-};
-
-/**
- * Describes the message zetachain.zetacore.crosschain.MsgUpdateERC20CustodyPauseStatusResponse.
- * Use `create(MsgUpdateERC20CustodyPauseStatusResponseSchema)` to create a new message.
- */
-export const MsgUpdateERC20CustodyPauseStatusResponseSchema: GenMessage<MsgUpdateERC20CustodyPauseStatusResponse> = /*@__PURE__*/
-  messageDesc(file_zetachain_zetacore_crosschain_tx, 29);
-
-/**
  * Msg defines the Msg service.
  *
  * @generated from service zetachain.zetacore.crosschain.Msg
@@ -1006,22 +905,6 @@ export const Msg: GenService<{
     methodKind: "unary";
     input: typeof MsgUpdateRateLimiterFlagsSchema;
     output: typeof MsgUpdateRateLimiterFlagsResponseSchema;
-  },
-  /**
-   * @generated from rpc zetachain.zetacore.crosschain.Msg.MigrateERC20CustodyFunds
-   */
-  migrateERC20CustodyFunds: {
-    methodKind: "unary";
-    input: typeof MsgMigrateERC20CustodyFundsSchema;
-    output: typeof MsgMigrateERC20CustodyFundsResponseSchema;
-  },
-  /**
-   * @generated from rpc zetachain.zetacore.crosschain.Msg.UpdateERC20CustodyPauseStatus
-   */
-  updateERC20CustodyPauseStatus: {
-    methodKind: "unary";
-    input: typeof MsgUpdateERC20CustodyPauseStatusSchema;
-    output: typeof MsgUpdateERC20CustodyPauseStatusResponseSchema;
   },
 }> = /*@__PURE__*/
   serviceDesc(file_zetachain_zetacore_crosschain_tx, 0);
